requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.2.1
groq==1.0.0
supabase==2.10.0
PyJWT==2.10.1